+++ core/src/main/java/jenkins/mvn/FilePathSettingsProvider.java
@@ -1,4 +1,4 @@
-package hudson.mvn;
+package jenkins.mvn;
 
 import hudson.Extension;
 import hudson.FilePath;
@@ -7,7 +7,6 @@ import hudson.model.TaskListener;
 import hudson.util.IOUtils;
 
 import java.io.File;
-import java.io.IOException;
 
 import org.apache.commons.lang.StringUtils;
 import org.kohsuke.stapler.DataBoundConstructor;
@@ -16,12 +15,12 @@ import org.kohsuke.stapler.DataBoundConstructor;
  * @author <a href="mailto:nicolas.deloof@gmail.com">Nicolas De Loof</a>
  * @author Dominik Bartholdi (imod)
  */
-public class FilePathGlobalSettingsProvider extends GlobalSettingsProvider {
+public class FilePathSettingsProvider extends SettingsProvider {
 
     private final String path;
 
     @DataBoundConstructor
-    public FilePathGlobalSettingsProvider(String path) {
+    public FilePathSettingsProvider(String path) {
         this.path = path;
     }
 
@@ -30,15 +29,14 @@ public class FilePathGlobalSettingsProvider extends GlobalSettingsProvider {
     }
 
     @Override
-    public FilePath supplySettings(AbstractBuild<?, ?> project, TaskListener listener) {
-        if (StringUtils.isEmpty(path)) {
+    public FilePath supplySettings(AbstractBuild<?, ?> build, TaskListener listener) {
+        if (StringUtils.isEmpty(path))
             return null;
-        }
         if (IOUtils.isAbsolute(path)) {
             return new FilePath(new File(path));
         } else {
-            FilePath mrSettings = project.getModuleRoot().child(path);
-            FilePath wsSettings = project.getWorkspace().child(path);
+            FilePath mrSettings = build.getModuleRoot().child(path);
+            FilePath wsSettings = build.getWorkspace().child(path);
             try {
                 if (!wsSettings.exists() && mrSettings.exists()) {
                     wsSettings = mrSettings;
@@ -52,12 +50,11 @@ public class FilePathGlobalSettingsProvider extends GlobalSettingsProvider {
     }
 
     @Extension(ordinal = 10)
-    public static class DescriptorImpl extends GlobalSettingsProviderDescriptor {
+    public static class DescriptorImpl extends SettingsProviderDescriptor {
 
         @Override
         public String getDisplayName() {
-            return "global settings file in project workspace";
+            return "File in project workspace";
         }
-
     }
 }
