+++ core/src/main/resources/jenkins/mvn/DefaultGlobalSettingsProvider/config.jelly
