+++ core/src/main/resources/jenkins/mvn/DefaultGlobalSettingsProvider/help.html
