+++ core/src/main/resources/jenkins/mvn/DefaultSettingsProvider/config.jelly
