+++ core/src/main/resources/jenkins/mvn/DefaultSettingsProvider/help.html
