+++ core/src/main/resources/jenkins/mvn/FilePathGlobalSettingsProvider/config.jelly
